renamed component
