+++ client/src/pages/presentational/FriendListDialog.jsx
